refactor(console): 중복 relTime → 공용 superadmin/relTime 로 통합
ErrorsPanel·SessionsPanel 에 동일하게 복제돼 있던 상대시간 포맷(방금/N분 전/
N시간 전/N일 전)을 relTime(ms) 한 곳으로 모음. SessionsPanel 은 ISO 라
호출부에서 new Date(iso).getTime() 으로 넘겨 동작 불변.
